requirements.txt
streamlit>=1.34.0
pandas>=2.2.0
numpy>=1.26.0
PyMuPDF>=1.24.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
python-dotenv>=1.0.1
openai>=1.35.0
ollama>=0.3.3
pytest>=8.2.0
